+++ expandedstorage/src/main/java/ninjaphenix/expandedstorage/common/block/CursedChestBlock.java
@@ -1,4 +1,4 @@
-package ninjaphenix.expandedstorage.common.content.block;
+package ninjaphenix.expandedstorage.common.block;
 
 import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
@@ -11,9 +11,9 @@ import net.minecraft.util.registry.SimpleRegistry;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 import ninjaphenix.expandedstorage.common.Registries;
-import ninjaphenix.expandedstorage.common.content.block.entity.CursedChestBlockEntity;
-import ninjaphenix.expandedstorage.common.content.misc.CursedChestType;
-import ninjaphenix.expandedstorage.common.content.ModContent;
+import ninjaphenix.expandedstorage.common.block.entity.CursedChestBlockEntity;
+import ninjaphenix.expandedstorage.common.misc.CursedChestType;
+import ninjaphenix.expandedstorage.common.ModContent;
 
 import static net.minecraft.state.property.Properties.HORIZONTAL_FACING;