+++ expandedstorage/src/main/java/ninjaphenix/expandedstorage/common/block/entity/CursedChestBlockEntity.java
@@ -1,4 +1,4 @@
-package ninjaphenix.expandedstorage.common.content.block.entity;
+package ninjaphenix.expandedstorage.common.block.entity;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.EnvironmentInterface;
@@ -20,9 +20,9 @@ import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import ninjaphenix.expandedstorage.common.Registries;
-import ninjaphenix.expandedstorage.common.content.block.BaseChestBlock;
-import ninjaphenix.expandedstorage.common.content.block.CursedChestBlock;
-import ninjaphenix.expandedstorage.common.content.ModContent;
+import ninjaphenix.expandedstorage.common.block.BaseChestBlock;
+import ninjaphenix.expandedstorage.common.block.CursedChestBlock;
+import ninjaphenix.expandedstorage.common.ModContent;
 import ninjaphenix.expandedstorage.common.inventory.AbstractScreenHandler;
 import ninjaphenix.expandedstorage.common.inventory.DoubleSidedInventory;