+++ expandedstorage/src/main/java/ninjaphenix/expandedstorage/mixin/ClientGameStart.java
@@ -1,7 +1,7 @@
 package ninjaphenix.expandedstorage.mixin;
 
 import net.minecraft.client.MinecraftClientGame;
-import ninjaphenix.expandedstorage.client.ContainerLibraryClient;
+import ninjaphenix.expandedstorage.client.ExpandedStorageClient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -11,5 +11,5 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class ClientGameStart
 {
     @Inject(method = "onStartGameSession", at = @At("TAIL"))
-    private void expandedstorage_onStartGameSession(final CallbackInfo ci) { ContainerLibraryClient.sendPreferencesToServer(); }
+    private void expandedstorage_onStartGameSession(final CallbackInfo ci) { ExpandedStorageClient.sendPreferencesToServer(); }
 }