+++ torcherino/changelog.md
@@ -1,3 +1,1 @@
-- Removed Minecraft and Fabric Loader dependencies due to constraint limitations
-  - Please use minecraft >= 1.16.2-rc1
-- Note: this build is intended for 1.16.2 however may not work as it is still targeting 1.16.2-rc2
+- Fixed crash at start up due to mixin refmap error.